+++ src/lib.rs
@@ -116,7 +116,7 @@ pub use ddl::{
 pub use functions::{
     abs, analysis_of_variance, any_last, any_value, arg_max, arg_min, array_concat, array_distinct,
     array_element, array_join, avg_if, avg_merge, avg_merge_state, avg_state, base64_decode,
-    base64_encode, ceil, city_hash64, concat, corr, cosine_distance, count_if, count_merge,
+    base64_encode, ceil, city_hash64, concat, corr, cosine_distance, count, count_if, count_merge,
     count_merge_state, count_state, covar_pop, covar_pop_stable, covar_samp, covar_samp_stable,
     cut_query_string, date_diff, date_trunc, domain, domain_without_www, empty, farm_fingerprint64,
     finalize_aggregation, first_significant_subdomain, floor, greatest, group_array,
@@ -155,8 +155,8 @@ pub use higher_order::{
 };
 pub use joins::{
     AliasedColumn, ClickHouseJoin, ClickHouseJoinBuilder, ClickHouseJoinDsl, JoinColumn, JoinKind,
-    JoinModifier, JoinOn, JoinStrictness, JoinUsing, clickhouse_join, join_column, source_column,
-    source_column_as,
+    JoinModifier, JoinOn, JoinStrictness, JoinUsing, clickhouse_join, expr_as, join_column,
+    source_column, source_column_as,
 };
 pub use json::{
     JsonPathFunction, JsonPathSegment, json_extract_bool_path, json_extract_float_path,